pyproject.toml
[tool.poetry]
name = "python-data-getter"
version = "0.1.0"
description = ""
authors = ["fernandoarielmateos@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
anthropic = "^0.45.2"
python-dotenv = "^1.0.1"
flask-marshmallow = "^1.3.0"
flask-sqlalchemy = "^3.1.1"
marshmallow-sqlalchemy = "^1.4.1"
flask-migrate = "^4.1.0"
flask-cors = "^5.0.0"
flask-fixtures = "^0.3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
src = ["app"]
fix = true
show-fixes = true
output-format = "full"

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # pyflakes
    "I",  # isort
    "UP",  # pyupgrade
    "W",  # pycodestyle warning
]

[tool.ruff.lint.isort]
force-single-line = true
order-by-type = false
